requirements.txt
us
django-crispy-forms
